# src/main/resources/openapi.yaml
openapi: "3.0.2"
info:
  title: API Ajmanre
  version: "1.0"
servers:
  - url: http://localhost:{port}
    description: The local API server
    variables:
      port:
        enum:
          - '8443'
          - '443'
        default: '8443'
  - url: https://angular.ajmanre.com:{port}
    description: The local API server
    variables:
      port:
        enum:
          - '8443'
          - '443'
        default: '8443'
...
paths:
...
  /prop/type:
    post:
      tags:
        - "propType"
      summary: "property type"
      operationId: "propType"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyTypeRequest"
      responses:
        "200":
          description: "MessageResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "405":
          description: "Invalid input"
...
  /agency/{agencyId}:
    get:
      tags:
        - "agencyId"
      summary: "agency get"
      operationId: "agencyId"
      parameters:
        - in: path
          name : agencyId
          schema:
            type: string
          required: true
          description: "agency details by id"
      responses:
        "200":
          description: "agency pagination listing"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agency"
...
  /agent:
    post:
      tags:
        - "agentAdd"
      summary: "add agent"
      operationId: "agentPost"
      parameters:
        - in: header
          name: Authorization
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentRequest"
      responses:
        "200":
          description: "MessageResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /agent/{page}/{size}:
    get:
      tags:
        - "agentPageSize"
      summary: "agent listing"
      operationId: "agentPageSize"
      parameters:
        - in: path
          name : page
          schema:
            type: integer
            default: 1
          required: true
          description: "page no of the list pagination"
        - in: path
          name : size
          schema:
            type: integer
            default: 5
          required: true
          description: "size of the pagination"
      responses:
        "200":
          description: "agency pagination listing"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentPage"
  /agent/{agentId}:
    get:
      tags:
        - "agentId"
      summary: "agent get"
      operationId: "agentId"
      parameters:
        - in: path
          name : agentId
          schema:
            type: string
          required: true
          description: "agent details by id"
      responses:
        "200":
          description: "agent pagination listing"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agent/{id}:
    put:
      tags:
        - "agentEdit"
      summary: "edit agent"
      operationId: "agentPut"
      parameters:
        - in: path
          name : id
          schema:
            type: string
          required: true
          description: "update agent details by id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Agent"
      responses:
        "200":
          description: "MessageResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /agent/by/{agencyId}:
    get:
      tags:
        - "agentByAgencyId"
      summary: "Agent By Agency Id"
      operationId: "agentByAgencyId"
      parameters:
        - in: path
          name : agencyId
          schema:
            type: string
          required: true
          description: "Agent By Agency Id"
      responses:
        "200":
          description: "Agent By Agency Id"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
...
  /enquiry:
    post:
      tags:
        - "enquiryAdd"
      summary: "add enquiry"
      operationId: "enquiryPost"
      parameters:
        - in: header
          name: Authorization
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Enquiry"
      responses:
        "200":
          description: "MessageResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /enquiry/{page}/{size}:
    get:
      tags:
        - "enquiryPageSize"
      summary: "enquiry listing"
      operationId: "enquiryPageSize"
      parameters:
        - in: path
          name : page
          schema:
            type: integer
            default: 1
          required: true
          description: "page no of the list pagination"
        - in: path
          name : size
          schema:
            type: integer
            default: 5
          required: true
          description: "size of the pagination"
      responses:
        "200":
          description: "lead pagination listing"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnquiryPage"
  /enquiry/{enquiryId}:
    get:
      tags:
        - "enquiryId"
      summary: "enquiry get"
      operationId: "enquiryId"
      parameters:
        - in: path
          name : enquiryId
          schema:
            type: string
          required: true
          description: "enquiry details by id"
      responses:
        "200":
          description: "enquiry pagination listing"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enquiry"
  /enquiry/{id}:
    put:
      tags:
        - "enquiryEdit"
      summary: "enquiry edit"
      operationId: "enquiryPut"
      parameters:
        - in: path
          name : id
          schema:
            type: string
          required: true
          description: "update enquiry details by id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Enquiry"
      responses:
        "200":
          description: "MessageResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: "Bad input"
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: "object"
      required:
        - message
      properties:
        message:
          type: "string"
          description: "Message"
        identifier:
          type: "string"
          description: "If identifier has to be in response"
    PropertyTypeRequest:
      type: "object"
      required:
        - type
        - order
      properties:
        id:
          type: "string"
          description: "hierarchy id"
        type:
          type: "string"
          description: "type of the property"
        order:
          type: "integer"
          description: "order list on this value"
        child:
          $ref: "#/components/schemas/PropertyTypeChild"
    PropertyTypeChild:
      type: "object"
      required:
        - type
        - order
      properties:
        type:
          type: "string"
          description: "type of the property"
        order:
          type: "integer"
          description: "order list on this value"
        child:
          $ref: "#/components/schemas/PropertyTypeChild"
...
    Address:
      type: "object"
      required:
        - adressLine1
        - city
      properties:
        adressLine1:
          type: "string"
        adressLine2:
          type: "string"
        area:
          type: "string"
        city:
          type: "string"
        country:
          type: "string"
        zip:
          type: "string"
    Contact:
      type: "object"
      required:
        - name
      properties:
        name:
          type: "string"
        tel:
          type: "string"
        mobile:
          type: "string"
        email:
          type: "string"
    Agency:
      type: "object"
      required:
        - id
        - name
      properties:
        id:
          type: "string"
        updatedAt:
          type: "string"
          format: "date-time"
        name:
          type: "string"
        tel:
          type: "string"
        facebook:
          type: "string"
          description: "facebook page"
        insta:
          type: "string"
          description: "instagram profile id"
        adresses:
          type: "array"
          description: "addresses of agency"
          items:
            $ref: "#/components/schemas/Address"
        contacts:
          type: "array"
          description: "contacts of agency"
          items:
            $ref: "#/components/schemas/Contact"
...
    AgentRequest:
      type: "object"
      required:
        - username
        - password
        - email
        - name
      properties:
        name:
          type: "string"
        serviceAreas:
          type: "string"
        specialties:
          type: "string"
          description: "agent specialties"
        position:
          type: "string"
          description: "agent position"
        license:
          type: "string"
          description: "agent position"
        imageFile:
          type: "string"
          description: "uploaded file url"
        address:
          $ref: "#/components/schemas/Address"
        contact:
          $ref: "#/components/schemas/Contact"
        agency:
          $ref: "#/components/schemas/Source"
        username:
          type: "string"
          description: "Username"
        password:
          type: "string"
          description: "Password"
        email:
          type: "string"
          description: "email"
    Agent:
      type: "object"
      required:
        - id
        - name
      properties:
        id:
          type: "string"
        updatedAt:
          type: "string"
          format: "date-time"
        name:
          type: "string"
        serviceAreas:
          type: "string"
        specialties:
          type: "string"
          description: "agent specialties"
        position:
          type: "string"
          description: "agent position"
        license:
          type: "string"
          description: "agent position"
        imageFile:
          type: "string"
          description: "uploaded file url"
        address:
          $ref: "#/components/schemas/Address"
        contact:
          $ref: "#/components/schemas/Contact"
        agency:
          $ref: "#/components/schemas/Source"
    AgentPage:
      type: "object"
      required:
        - page
        - size
        - total
      properties:
        page:
          type: integer
          description: page no of data result in pagination
        size:
          type: integer
          description: no of data elements in page
        data:
          type: array
          items:
            $ref: "#/components/schemas/Agent"
          description: data elements list in page
        total:
          type: integer
          format: int64
          description: total elements in the list
    Source:
      type: "object"
      required:
        - id
        - name
      properties:
        id:
          type: string
          description: source identifier
        name:
          type: string
          description: key value of the  source
...
    Enquiry:
      type: "object"
      required:
        - contact
        - propType
        - propTypeId
        - enqType
        - enqTypeId
      properties:
        id:
          type: string
          description: source identifier
        updatedAt:
          type: "string"
          format: "date-time"
        contact:
          $ref: "#/components/schemas/Contact"
        enqType:
          type: string
        enqTypeId:
          type: string
        propType:
          type: string
        propTypeId:
          type: string
        priceFrom:
          type: number
          format: double
        priceTo:
          type: number
          format: double
        bedMin:
          type: integer
        bedMax:
          type: integer
        bathMin:
          type: integer
        bathMax:
          type: integer
        sqftMin:
          type: number
          format: double
        sqftMax:
          type: number
          format: double
        address:
          $ref: "#/components/schemas/Address"
        notes:
          type: string
          description: notes for the person who is inquiring
    EnquiryPage:
      type: "object"
      required:
        - page
        - size
        - total
      properties:
        page:
          type: integer
          description: page no of data result in pagination
        size:
          type: integer
          description: no of data elements in page
        data:
          type: array
          items:
            $ref: "#/components/schemas/Enquiry"
          description: data elements list in page
        total:
          type: integer
          format: int64
          description: total elements in the list
security:
  - bearerAuth: []  